# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.23.0"
  },
  "paths": {
...
    "/agents/builtin/{name}/extensions/bind": {
      "post": {
        "tags": [
          "Builtin Agents"
        ],
        "operationId": "bind_extension_to_agent",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Agent name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindExtensionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Extension bound"
          }
        }
      }
    },
    "/agents/builtin/{name}/extensions/unbind": {
      "post": {
        "tags": [
          "Builtin Agents"
        ],
        "operationId": "unbind_extension_from_agent",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Agent name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindExtensionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Extension unbound"
          }
        }
      }
    },
    "/agents/builtin/{name}/toggle": {
      "post": {
        "tags": [
          "Builtin Agents"
        ],
        "operationId": "toggle_builtin_agent",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Agent name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToggleAgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
...
    "/agents/external/connect": {
      "post": {
        "tags": [
          "External Agents"
        ],
        "operationId": "connect_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectAgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found"
          },
          "422": {
            "description": "Agent has no distribution"
          }
        }
      }
    },
    "/agents/external/{agent_id}": {
      "delete": {
        "tags": [
          "External Agents"
        ],
        "operationId": "disconnect_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "description": "Agent identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent disconnected"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agents/external/{agent_id}/mode": {
      "post": {
        "tags": [
          "External Agents"
        ],
        "operationId": "set_mode",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "description": "Agent identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetModeAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Mode set"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agents/external/{agent_id}/prompt": {
      "post": {
        "tags": [
          "External Agents"
        ],
        "operationId": "prompt_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "description": "Agent identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Prompt response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptAgentResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agents/external/{agent_id}/session": {
      "post": {
        "tags": [
          "External Agents"
        ],
        "operationId": "create_session",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "description": "Agent identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BindExtensionRequest": {
        "type": "object",
        "required": [
          "extension_name"
        ],
        "properties": {
          "extension_name": {
            "type": "string"
          }
        }
      },
...
      "ConnectAgentRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ConnectAgentResponse": {
        "type": "object",
        "required": [
          "agent_id",
          "connected"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "connected": {
            "type": "boolean"
          }
        }
      },
...
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "working_dir": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "PromptAgentRequest": {
        "type": "object",
        "required": [
          "session_id",
          "text"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "PromptAgentResponse": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
...
      "SetModeAgentRequest": {
        "type": "object",
        "required": [
          "session_id",
          "mode_id"
        ],
        "properties": {
          "mode_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "ToggleAgentResponse": {
        "type": "object",
        "required": [
          "name",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}